requirements.txt
flask==3.0.2
flask-cors==4.0.0
numpy<2
transformers==4.39.3
torch==2.2.1
python-dotenv==1.0.1
huggingface_hub[hf_xet]==0.23.1
google-generativeai==0.3.2
